requirements.txt
pyradiomics
fastai
pydot
pandas
numpy
matplotlib
sklearn
datetime
tqdm
pynrrd
PIL